Ensure that the base URL always updates the configuration
